+++ src/main/java/com/epam/msu/entity/Certificate.java
@@ -9,9 +9,7 @@ import java.util.Objects;
 
 public class Certificate {
 
-    //MapStruct
-    //LocalDateTime
-    private long id;
+    private int id;
     private String name;
     private String description;
     private long price;
@@ -22,7 +20,7 @@ public class Certificate {
     public Certificate() {
     }
 
-    public Certificate(long id, String name, String description, long price, int duration, Timestamp createDate, Timestamp lastUpdateDate) {
+    public Certificate(int id, String name, String description, long price, int duration, Timestamp createDate, Timestamp lastUpdateDate) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -31,11 +29,11 @@ public class Certificate {
         this.createDate = createDate;
         this.lastUpdateDate = lastUpdateDate;
     }
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
