+++ src/main/java/com/epam/msu/entity/Tag.java
@@ -4,12 +4,12 @@ import java.util.Objects;
 
 public class Tag {
 
-    private long id;
+    private int id;
     private String name;
 
     public Tag(){}
 
-    public Tag(long id, String name) {
+    public Tag(int id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -18,11 +18,11 @@ public class Tag {
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
